@@ -1,7 +1,8 @@
 from datetime import date
+from enum import Enum
 
 import strawberry
-from enum import Enum
+
 
 @strawberry.input
 class DateFilter:
@@ -9,11 +10,13 @@ class DateFilter:
     items: list[date] | None = None
     year: int | None = None
 
+
 @strawberry.enum
 class Order(Enum):
     asc = "asc"
     desc = "desc"
-    
+
+
 @strawberry.input
 class SortModel:
     field: str
